Detect whether terminal-notifier or notify-send are available on startup.
This way, the shell is not looking for the command every single time a message
is received from the ØMQ socket. While the frequency of received messages is
pretty much guaranteed to be low enough that this won't actually affect the
performance of the script, I think this is a cleaner way of handling the
different notifier programs.

The only downside of this is that if the user removes the notifier program while
the client is running it will spew error messages to stdout, but I really think
that should be the user's concern and not ours.