Fix messages/nicks that begin with [, {, (, < for terminal-notifier.
terminal-notifier's help text says "Note that in some circumstances the first
character of a message has to be escaped in order to be recognized." Trial and
error testing revealed that the four opening brackets listed above needed to be
escaped to be recognized properly. This was by no means exhaustive testing and
will not necessarily fix all cases, but should be good enough for the most
common ones.

Possibly of note is that [ and { are valid characters for beginning a nick,
while < and ( are not even allowed in a nick. Regardless of the fact that a nick
cannot start with ( or <, the same regular expression is used to fix both the
message and the nick.